improved documentation
