+++ accessories/heroku/Procfile
@@ -1,0 +1,1 @@
+web: gunicorn  mzqc_heroku_validator:app
