+++ accessories/heroku/requirements.txt
@@ -1,0 +1,81 @@
+pymzqc
+#git+git://github.com/MS-Quality-hub/pymzqc.git@v1.0.0
+# aniso8601==9.0.1
+# argon2-cffi==20.1.0
+# async-generator==1.10
+# attrs==21.2.0
+# backcall==0.2.0
+# bleach==3.3.1
+# certifi==2021.5.30
+# cffi==1.14.6
+# chardet==4.0.0
+# charset-normalizer==2.0.3
+# click==8.0.1
+# cycler==0.10.0
+# debugpy==1.3.0
+# decorator==4.4.2
+# defusedxml==0.7.1
+# entrypoints==0.3
+# fastobo==0.10.1
+Flask==2.0.1
+Flask-Cors==3.0.10
+Flask-RESTful==0.3.9
+gunicorn==20.1.0
+idna==3.2
+ipykernel==6.0.3
+ipython==7.31.1
+ipython-genutils==0.2.0
+ipywidgets==7.6.3
+itsdangerous==2.0.1
+jedi==0.18.0
+Jinja2==3.0.1
+jsonschema==3.2.0
+jupyter==1.0.0
+jupyter-client==6.1.12
+jupyter-console==6.4.0
+jupyter-core==4.7.1
+jupyterlab-pygments==0.1.2
+jupyterlab-widgets==1.0.0
+kiwisolver==1.3.1
+MarkupSafe==2.0.1
+matplotlib==3.4.2
+matplotlib-inline==0.1.2
+mistune==0.8.4
+nbclient==0.5.3
+nbconvert==6.1.0
+nbformat==5.1.3
+nest-asyncio==1.5.1
+networkx==2.5.1
+notebook==6.4.12
+numpy==1.22.0
+packaging==21.0
+pandocfilters==1.4.3
+parso==0.8.2
+pexpect==4.8.0
+pickleshare==0.7.5
+Pillow==9.0.1
+prometheus-client==0.11.0
+prompt-toolkit==3.0.19
+pronto==2.4.2
+ptyprocess==0.7.0
+pycparser==2.20
+Pygments==2.9.0
+pyparsing==2.4.7
+pyrsistent==0.18.0
+python-dateutil==2.8.2
+pytz==2021.1
+pyzmq==22.1.0
+qtconsole==5.1.1
+QtPy==1.9.0
+requests==2.26.0
+Send2Trash==1.7.1
+six==1.16.0
+terminado==0.10.1
+testpath==0.5.0
+tornado==6.1
+traitlets==5.0.5
+urllib3==1.26.6
+wcwidth==0.2.5
+webencodings==0.5.1
+Werkzeug==2.0.1
+widgetsnbextension==3.5.1
