+++ accessories/heroku/wsgi.py
@@ -1,0 +1,4 @@
+from mzqc_heroku_validator import app
+
+if __name__ == '__main__':
+    app.run()
