--- acessories/heroku/runtime.txt
@@ -1,1 +1,0 @@
-python-3.8.11
