+++ containers/README.md
@@ -1,0 +1,9 @@
+# Containers
+
+This folder contains various container build instruction files.
+The `DevDockerfile` builds a container from github sources and installs some helpful modules before that.
+The `Dockerfile` builds a container from a wheel build.
+More container for development have their instructions in the /.devcontainer folder
+
+## Repository
+Our main container repository is at [quay.io](https://quay.io/repository/mwalzer/pymzqc?tab=tags)
