+++ doc/source/_static/css/pymzqc_theme.css
@@ -1,0 +1,5 @@
+@import url("theme.css");
+
+.wy-nav-content {
+    max-width: 80%;
+}
