+++ doc/source/accessories.rst
@@ -1,0 +1,8 @@
+Accessories
+============
+
+The accessories currently include the following subprojects:
+
+Validator API (heroku)
+----------
+.. include:: ../../accessories/heroku/README.md
