+++ doc/source/containers.rst
@@ -1,0 +1,4 @@
+mzqc package
+============
+
+.. include:: ../../containers/README.md
