+++ doc/source/examples.rst
@@ -2,6 +2,7 @@ Usage examples
 **************
 I am a fan of quick hands-on overviews to get to know software, that I might or might not use in the future.
 Several hands-on examples can be found `here. <https://github.com/MS-Quality-hub/pymzqc/tree/main/jupyter>`_ 
+Simply open them from `colab. <https://colab.research.google.com/>`_ 
 A video version for ASMS'22 is `on youtube: <https://www.youtube.com/watch?v=vZXJuPl2yGw>`_
 
 The copy&paste essentials:
