+++ doc/source/notebooks.rst
@@ -1,0 +1,14 @@
+Python notebooks
+============
+
+.. include:: ../../jupyter/README.md
+    
+colab write
+--------------------
+
+.. include:: ../../jupyter/colab/write_in_5_minutes.md
+
+colab read
+--------------------
+
+.. include:: ../../jupyter/colab/read_in_5_minutes.md
