+++ jupyter/README.md
@@ -1,0 +1,12 @@
+## Jupyter python notebooks
+
+This folder contains various iterations of python notebooks for different occasions.
+We can recommend the colab notebooks:
+* [read mzQC in 5 minutes](colab/read_in_5_minutes.md)
+* [write mzQC in 5 minutes](colab/write_in_5_minutes.md)
+
+### preview update 
+```
+jupyter nbconvert --to markdown '/workspaces/pymzqc/jupyter/colab/write_in_5_minutes.ipynb'
+```
+This will create an updated static markdown render to /workspaces/pymzqc/jupyter/colab/write_in_5_minutes.md
